feat(mount_drive): add script to mount drives on linux
This commit adds a script to mount drives on linux. The script prompts the user to enter the device name and mount name. If the user does not enter any value, the script sets the default values. The script checks if the drive is bitlocker encrypted and installs dislocker if it is not installed. If the drive is bitlocker encrypted, the script creates a directory for the bitlocker mount and mounts the drive using dislocker. If the drive is not encrypted, the script mounts the drive normally.